codewars-II/requirements.txt
autopep8==1.6.0
mysql-connector-python==8.0.29
mysqlclient==2.1.1
protobuf==4.21.2
pycodestyle==2.8.0
python-dotenv==0.20.0
toml==0.10.2